chore: remove @allanoricil/node-red-node
@allanoricil/node-red-node is now a peer dependency of the nrg cli
I will add node-red as well, so that projects don't need to have them
defined.

define minimum versions for node and npm

add nrg as a keyword because this project is now using nrg
remove esbuild because it is baked inside nrg, and in the future I
plan to enable rollup too.
add esm to refer to the module system used to write the code
add es2022 to refer to the spec of the build