chore(deps): bump @allanoricil/nrg-cli from 0.0.0-beta.7 to 0.0.0-beta.8
This version comes with a proper logger. The default log level is "info"
for console messages, and "silly" for logs stored in a file.

Logs are created in a folder in the root of the project called logs/nrg.log

Log level can be set using the following env variable

NRG_LOG_LEVEL='error' | 'warn' | 'info' | 'debug' | 'verbose' | 'silly'